requirements.txt
streamlit
opencv-python
numpy
Pillow
moviepy
torch
ultralytics